requirements.txt
pillow
tkinter
